refactor: align mobile diff package boundaries
