+++ Packages/iOS/CmuxMobileBrowser/Sources/CmuxMobileBrowser/MobileDiffPatchContent.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct MobileDiffPatchContent: Sendable {
+    let data: Data
+    let mimeType: String
+}
