+++ Packages/iOS/CmuxMobileBrowser/Sources/CmuxMobileBrowser/MobileDiffPatchPayload.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct MobileDiffPatchPayload: Sendable {
+    let html: Data
+    let patch: Data
+}
